Allow for integration tests in tests/ and cleanup the project setup.
  - Cleanup the usage of “extern crate” by gathering them in lib.rs.
  - Added the library crate “rush”.
  - Move tests into tests/.
