+++ Cargo.toml
@@ -10,6 +10,9 @@ edition = "2018"
 [features]
 nightly = []
 
+[lib]
+name = "rush"
+
 [[bin]]
 name = "rush"
 path = "src/main.rs"
