+++ src/variables.rs
@@ -19,15 +19,13 @@
 // MA 02110-1301, USA.
 //
 
-/// RuSh variables management is located in this module.
-///
-/// variables.rs contains variables structures and affiliated methods.
-/// `Variable` and `Variables` are defined here.
-/// variables (un)setting, update methods for classical variables.
-extern crate rand;
-extern crate seahash;
+//! RuSh variables management is located in this module.
+//!
+//! variables.rs contains variables structures and affiliated methods.
+//! `Variable` and `Variables` are defined here.
+//! variables (un)setting, update methods for classical variables.
 
-use self::rand::Rng;
+use rand::Rng;
 use libc::{c_char, c_int, geteuid, getgid, getlogin, getpid, getppid, getuid, size_t};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::HashMap;
@@ -79,7 +77,7 @@ impl Variable {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Value};
+    /// use rush::variables::{Access, Variable, Value};
     /// let var = Variable { value: Value::I(-42), access: Access::ReadWrite };
     /// assert_eq!(var.geti(), -42);
     /// ```
@@ -94,7 +92,7 @@ impl Variable {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Value};
+    /// use rush::variables::{Access, Variable, Value};
     /// let var = Variable { value: Value::F(-42.5), access: Access::ReadWrite };
     /// assert_eq!(var.getf(), -42.5);
     /// ```
@@ -109,7 +107,7 @@ impl Variable {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Value};
+    /// use rush::variables::{Access, Variable, Value};
     /// let var = Variable { value: Value::S("Forty two".to_string()), access: Access::ReadWrite };
     /// assert_eq!(var.gets(), "Forty two");
     /// ```
@@ -133,7 +131,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Variables, Value};
+    /// use rush::variables::{Access, Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// match vars.get("RUSH_COMMAND") {
@@ -167,7 +165,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Variables, Value};
+    /// use rush::variables::{Access, Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// match vars.get_access("RUSH_COMMAND") {
@@ -194,7 +192,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Variables, Value};
+    /// use rush::variables::{Access, Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// vars.set(String::from("TESTF"), Variable { value: Value::F(-49.3), access: Access::ReadWrite });
@@ -232,7 +230,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Variables, Value};
+    /// use rush::variables::{Access, Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// vars.set_access("TEST".to_string(), Access::ReadWrite);
@@ -273,7 +271,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Variable, Variables, Value};
+    /// use rush::variables::{Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// match vars.get("RUSH_COMMAND") {
@@ -295,7 +293,7 @@ impl Variables {
     ///
     /// # Examples
     /// ```rust
-    /// use RuSh::variables::{Access, Variable, Variables, Value};
+    /// use rush::variables::{Access, Variable, Variables, Value};
     ///
     /// let mut vars = Variables::init_shell_vars();
     /// match vars.get("RUSH") {
@@ -596,142 +594,5 @@ impl Variables {
             },
         );
         vars
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    //use crate::variables::Variables;
-    use crate::variables::{Access, Value, Variable, Variables};
-
-    #[test]
-    fn test_init_shell_vars() {
-        let vars = Variables::init_shell_vars();
-        match vars.get("RUSH_COMMAND") {
-            Some(v) => assert_eq!(v.gets(), ""),
-            None => panic!("RUSH_COMMAND should be defined."),
-        }
-        match vars.get("HISTSIZE") {
-            Some(v) => assert_eq!(v.geti(), 1000),
-            None => panic!("HISTSIZE should be defined."),
-        }
-    }
-
-    #[test]
-    fn test_unset() {
-        let mut vars = Variables::init_shell_vars();
-        match vars.get("RUSH_COMMAND") {
-            Some(v) => assert_eq!(v.gets(), ""),
-            None => panic!("RUSH_COMMAND should be defined."),
-        }
-        vars.unset(String::from("RUSH_COMMAND"));
-        match vars.get("RUSH_COMMAND") {
-            Some(_v) => panic!("RUSH_COMMAND should have been unset."),
-            None => println!("RUSH_COMMAND is not set."),
-        }
-    }
-
-    #[test]
-    fn test_get_and_getifs() {
-        let mut vars = Variables::init_shell_vars();
-        match vars.get("RUSH_COMMAND") {
-            Some(v) => assert_eq!(v.gets(), ""),
-            None => panic!("RUSH_COMMAND should be defined."),
-        }
-        match vars.get("HISTSIZE") {
-            Some(v) => assert_eq!(v.geti(), 1000),
-            None => panic!("HISTSIZE should be defined."),
-        }
-        vars.set(
-            String::from("TEST"),
-            Variable {
-                value: Value::F(-49.3),
-                access: Access::ReadWrite,
-            },
-        );
-        match vars.get("TEST") {
-            Some(v) => assert_eq!(v.getf(), -49.3),
-            None => panic!("TEST variable should be defined."),
-        }
-    }
-
-    #[test]
-    fn test_set() {
-        let mut vars = Variables::init_shell_vars();
-        vars.set(
-            String::from("TESTF"),
-            Variable {
-                value: Value::F(-49.3),
-                access: Access::ReadWrite,
-            },
-        );
-        match vars.get("TESTF") {
-            Some(v) => assert_eq!(v.getf(), -49.3),
-            None => panic!("TESTF should be defined."),
-        }
-        vars.set(
-            String::from("TESTI"),
-            Variable {
-                value: Value::I(-42),
-                access: Access::ReadWrite,
-            },
-        );
-        match vars.get("TESTI") {
-            Some(v) => assert_eq!(v.geti(), -42),
-            None => panic!("TESTI should be defined."),
-        }
-        vars.set(
-            String::from("TESTS"),
-            Variable {
-                value: Value::S(String::from("RuSh will rock (one day)")),
-                access: Access::ReadWrite,
-            },
-        );
-        match vars.get("TESTS") {
-            Some(v) => assert_eq!(v.gets(), "RuSh will rock (one day)"),
-            None => panic!("TESTS variable should be defined."),
-        }
-    }
-
-    #[test]
-    fn test_get_access() {
-        let vars = Variables::init_shell_vars();
-        match vars.get_access("RUSH_COMMAND") {
-            Some(v) => assert_eq!(v, Access::ReadWrite),
-            None => panic!("RUSH_COMMAND should be defined and Access::ReadWrite."),
-        }
-        match vars.get_access("EUID") {
-            Some(v) => assert_eq!(v, Access::ReadOnly),
-            None => panic!("EUID should be defined and Access::ReadOnly."),
-        }
-        match vars.get_access("nonexistingvar") {
-            Some(v) => panic!("nonexistingvar should not give back {:?}", v),
-            None => assert!(true),
-        }
-    }
-
-    #[test]
-    fn test_set_access() {
-        let mut vars = Variables::init_shell_vars();
-        vars.set_access("TEST".to_string(), Access::ReadWrite);
-        match vars.get_access("TEST") {
-            Some(v) => assert_eq!(v, Access::ReadWrite),
-            None => panic!("TEST should be defined."),
-        }
-        vars.set_access("TEST".to_string(), Access::ReadOnly);
-        match vars.get_access("TEST") {
-            Some(v) => assert_eq!(v, Access::ReadOnly),
-            None => panic!("TEST should be defined."),
-        }
-        vars.set_access("doesnotexist".to_string(), Access::ReadWrite);
-        match vars.get_access("doesnotexist") {
-            Some(v) => assert_eq!(v, Access::ReadWrite),
-            None => panic!("doesnotexist variable should be defined and Access::ReadWrite"),
-        }
-        vars.set_access("doesnotexist2".to_string(), Access::ReadOnly);
-        match vars.get_access("doesnotexist2") {
-            Some(v) => assert_eq!(v, Access::ReadOnly),
-            None => panic!("doesnotexist variable should be defined and Access::ReadOnly"),
-        }
     }
 }
